+++ CHANGELOG.rst
@@ -1,3 +1,10 @@
+Development version
+-------------------
+
+* fixed pyws crashing if SOAP request body does not match the required schema,
+issue #16.
+
+
 Version 1.1
 -----------